requirements.txt
# Core Flask Dependencies
Flask==2.2.2
flask-cors==3.0.10
werkzeug==2.2.2
gunicorn==20.1.0

# Phase 1 Enhancements
python-dotenv==1.0.0
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
Flask-Compress==1.14

# Phase 2 Enhancements
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
PyJWT==2.8.0
bcrypt==4.1.2

# AI/ML Dependencies
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==9.5.0
matplotlib==3.7.1

# Utilities
tqdm==4.65.0
typing-extensions==4.5.0

# Testing (for Phase 1)
pytest==7.4.0
pytest-flask==1.2.0
